+++ ewm-stats-service/stats-service/src/main/java/ru/practicum/repository/EndpointHitRepository.java
@@ -11,9 +11,7 @@ import java.util.Collection;
 import java.util.List;
 
 @Repository
-public interface EndpointHitRepository extends JpaRepository<EndpointHit, Long> {
-
-    // Получить статистику для всех uri без учёта уникальных ip (uri == null, unique == false)
+public interface EndpointHitRepository extends JpaRepository<EndpointHit, Integer> {
 
     @Query("SELECT new ru.practicum.model.ViewStats(eh.app, eh.uri, COUNT(eh.id)) " +
            "FROM EndpointHit as eh " +
@@ -22,16 +20,12 @@ public interface EndpointHitRepository extends JpaRepository<EndpointHit, Long> 
            "ORDER BY COUNT(eh.id) DESC")
     List<ViewStats> getAllServicesStats(LocalDateTime startDateTime, LocalDateTime endDateTime);
 
-    // Получить статистику для всех uri с учётом уникальных ip (uri == null, unique == true)
-
     @Query("SELECT new ru.practicum.model.ViewStats(eh.app, eh.uri, COUNT(DISTINCT (eh.ip))) " +
            "FROM EndpointHit as eh " +
            "WHERE (eh.timestamp BETWEEN :startDateTime AND :endDateTime) " +
            "GROUP BY eh.app, eh.uri " +
            "ORDER BY COUNT(DISTINCT (eh.ip)) DESC")
     List<ViewStats> getAllServicesStatsWithUniqueIp(LocalDateTime startDateTime, LocalDateTime endDateTime);
-
-    // Получить статистику для заданных uri без учёта уникальных ip (uri != null, unique == false)
 
     @Query("SELECT new ru.practicum.model.ViewStats(eh.app, eh.uri, COUNT(eh.id)) " +
            "FROM EndpointHit as eh " +
@@ -40,8 +34,6 @@ public interface EndpointHitRepository extends JpaRepository<EndpointHit, Long> 
            "GROUP BY eh.app, eh.uri " +
            "ORDER BY COUNT(eh.id) DESC")
     List<ViewStats> getStats(LocalDateTime startDateTime, LocalDateTime endDateTime, Collection<String> uris);
-
-    // Получить статистику для заданных uri с учётом уникальных ip (uri != null, unique == true)
 
     @Query("SELECT new ru.practicum.model.ViewStats(eh.app, eh.uri, COUNT(DISTINCT (eh.ip))) " +
            "FROM EndpointHit as eh " +
